requirements.txt
faiss-cpu==1.8.0.post1
google-ai-generativelanguage==0.6.6
google-auth==2.34.0
google-generativeai==0.7.2
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
pypdf==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.38.0